api/requirements.txt
Django>=3.2.7,<4.0.0
djangorestframework>=3.12.4,<4.0.0
django-cors-headers>=3.8.0,<4.0.0
django-filter>=2.4.0,<3.0.0
psycopg2>=2.9.1,<3.0.0
Pillow>=8.3.2,<9.0.0
django-computed-property>=0.3.0,<1.0.0
djangorestframework_jwt>=1.11.0,<2.0.0
git+https://github.com/Peter-Slump/django-keycloak.git#egg=django-keycloak

flake8>=3.9.2,<4.0.0